+++ db/migrate/20151026192731_add_chair_to_committee_members.rb
@@ -1,0 +1,5 @@
+class AddChairToCommitteeMembers < ActiveRecord::Migration
+  def change
+    add_column :committee_members, :chair, :boolean, null: false, default: false
+  end
+end